+++ core/src/com/visualg/animations/fractals/multibrot/MultibrotAlg.java
@@ -34,8 +34,8 @@ class MultibrotAlg extends PixmapAlg {
         for (int x = 0; x < WIDTH; x++) {
             for (int y = 0; y < HEIGHT; y++) {
                 updateProgress(x, y);
-                double re = resolutionShapeAdjustment * (x - (double) WIDTH / 2) / (WIDTH * 0.5 * zoom) - xOffset;
-                double im = (y - (double) HEIGHT / 2) / (HEIGHT * 0.5 * zoom) - yOffset;
+                double re = resolutionShapeAdjustment * (x - WIDTH / 2d) / (WIDTH * 0.5 * zoom) - xOffset;
+                double im = (y - HEIGHT / 2d) / (HEIGHT * 0.5 * zoom) - yOffset;
                 double prevRe = 0;
                 double prevIm = 0;
                 double nextRe, nextIm;
